Do not remove future txns
Collect invalid future txns and put them in a cache. Stop removing the detected invalid future txns when finalizing blocks. Check invalid future txns for miner when it checks the txn confirmation. With the changes above, we can avoid changing the protocol that remove future txns in server end, but also keep the fast invalid txn detecting and retry.